move check member has role method to roles file for consistency
