+++ ironforgedbot/common/helpers.py
@@ -1,15 +1,14 @@
 import logging
+import re
 from datetime import datetime
 from io import BytesIO
-import re
-from typing import List, Optional, Tuple, TypedDict
+from typing import List, Tuple, TypedDict
 
 import discord
 from dateutil.relativedelta import relativedelta
 from discord import Guild, Member
 
 from ironforgedbot.common.constants import MAX_DISCORD_MESSAGE_SIZE, NEW_LINE, QUOTES
-from ironforgedbot.common.roles import ROLE
 from ironforgedbot.http import HTTP
 
 logger = logging.getLogger(__name__)
@@ -74,30 +73,6 @@ def validate_playername(
         return find_member_by_nickname(guild, playername), playername
     except ValueError:
         return None, playername
-
-
-def check_member_has_role(
-    member: Member,
-    required_role: ROLE,
-    or_higher: Optional[bool] = False,
-    or_lower: Optional[bool] = False,
-) -> bool:
-    member_roles = set(role.name for role in member.roles)
-    acceptable_roles = set([required_role.value])
-
-    if or_higher:
-        acceptable_roles = set(ROLE(required_role).or_higher())
-
-    if or_lower:
-        acceptable_roles = set(ROLE(required_role).or_lower())
-
-    member_roles = {role.lower().strip() for role in member_roles}
-    acceptable_roles = {role.lower().strip() for role in acceptable_roles}
-
-    if acceptable_roles & member_roles:
-        return True
-
-    return False
 
 
 def find_member_by_nickname(guild: Guild, target_name: str) -> Member:
