+++ tests/common/helpers_test.py
@@ -10,10 +10,9 @@ from ironforgedbot.common.helpers import (
     normalize_discord_string,
     render_percentage,
     render_relative_time,
-    check_member_has_role,
     validate_playername,
 )
-from ironforgedbot.common.roles import ROLE
+from ironforgedbot.common.roles import ROLE, check_member_has_role
 from tests.helpers import create_mock_discord_interaction, create_test_member
 
 
@@ -105,56 +104,6 @@ class TestHelpers(unittest.TestCase):
 
         self.assertEqual(result_member, None)
         self.assertEqual(result_playername, playername)
-
-    def test_check_member_has_role(self):
-        """Test validate member has role"""
-        member = Mock(discord.Member)
-        member.roles = [ROLE.MEMBER]
-
-        self.assertEqual(check_member_has_role(member, ROLE.MEMBER), True)
-
-    def test_check_member_has_role_fails(self):
-        """Test validate member has role fails when member does not have role"""
-        member = Mock(discord.Member)
-        member.roles = [ROLE.MEMBER]
-
-        self.assertEqual(check_member_has_role(member, ROLE.LEADERSHIP), False)
-
-    def test_check_member_has_role_or_higher(self):
-        """Test validate member has role or higher"""
-        member = Mock(discord.Member)
-        member.roles = [ROLE.LEADERSHIP]
-
-        self.assertEqual(
-            check_member_has_role(member, ROLE.MEMBER, or_higher=True), True
-        )
-
-    def test_check_member_has_role_or_higher_fail(self):
-        """Test validate member has role or higher"""
-        member = Mock(discord.Member)
-        member.roles = [ROLE.MEMBER]
-
-        self.assertEqual(
-            check_member_has_role(member, ROLE.STAFF, or_higher=True), False
-        )
-
-    def test_check_member_has_role_or_lower(self):
-        """Test validate member has role or lower"""
-        member = Mock(discord.Member)
-        member.roles = [ROLE.MEMBER]
-
-        self.assertEqual(
-            check_member_has_role(member, ROLE.LEADERSHIP, or_lower=True), True
-        )
-
-    def test_check_member_has_role_or_lower_fail(self):
-        """Test validate member has role or lower"""
-        member = Mock(discord.Member)
-        member.roles = []
-
-        self.assertEqual(
-            check_member_has_role(member, ROLE.MEMBER, or_lower=True), False
-        )
 
     def test_find_member_by_nickname(self):
         """Test find member by nickname happy path"""
