+++ tests/common/roles_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from ironforgedbot.common.roles import ROLE, is_member, is_prospect
+from ironforgedbot.common.roles import ROLE, check_member_has_role
 from tests.helpers import create_test_member
 
 
@@ -17,14 +17,45 @@ class TestRoles(unittest.TestCase):
 
         self.assertEqual(expected, result)
 
-    def test_is_prospect(self):
-        prospect_member = create_test_member("test", ["Member", ROLE.PROSPECT])
-        member = create_test_member("test", ["Member", ROLE.LEADERSHIP, "Tester"])
-        self.assertEqual(is_prospect(prospect_member), True)
-        self.assertEqual(is_prospect(member), False)
+    def test_check_member_has_role(self):
+        """Test validate member has role"""
+        member = create_test_member("", [ROLE.MEMBER])
+        self.assertEqual(check_member_has_role(member, ROLE.MEMBER), True)
 
-    def test_is_member(self):
-        member = create_test_member("test", ["Tester", ROLE.MEMBER])
-        non_member = create_test_member("test", ["Tester", "Foo", "Bar"])
-        self.assertEqual(is_member(member), True)
-        self.assertEqual(is_member(non_member), False)
+    def test_check_member_has_role_fails(self):
+        """Test validate member has role fails when member does not have role"""
+        member = create_test_member("", [ROLE.MEMBER])
+
+        self.assertEqual(check_member_has_role(member, ROLE.LEADERSHIP), False)
+
+    def test_check_member_has_role_or_higher(self):
+        """Test validate member has role or higher"""
+        member = create_test_member("", [ROLE.LEADERSHIP])
+
+        self.assertEqual(
+            check_member_has_role(member, ROLE.MEMBER, or_higher=True), True
+        )
+
+    def test_check_member_has_role_or_higher_fail(self):
+        """Test validate member has role or higher"""
+        member = create_test_member("", [ROLE.MEMBER])
+
+        self.assertEqual(
+            check_member_has_role(member, ROLE.STAFF, or_higher=True), False
+        )
+
+    def test_check_member_has_role_or_lower(self):
+        """Test validate member has role or lower"""
+        member = create_test_member("", [ROLE.MEMBER])
+
+        self.assertEqual(
+            check_member_has_role(member, ROLE.LEADERSHIP, or_lower=True), True
+        )
+
+    def test_check_member_has_role_or_lower_fail(self):
+        """Test validate member has role or lower"""
+        member = create_test_member("", [])
+
+        self.assertEqual(
+            check_member_has_role(member, ROLE.MEMBER, or_lower=True), False
+        )
